move: package 명 변경 (chessBoard -> chessgame)
