+++ src/main/java/chess/domain/chessGame/FileScore.java
@@ -1,4 +1,4 @@
-package chess.domain.chessBoard;
+package chess.domain.chessGame;
 
 import chess.domain.position.File;
 import java.util.ArrayList;
