+++ src/main/java/chess/domain/chessGame/Space.java
@@ -1,4 +1,4 @@
-package chess.domain.chessBoard;
+package chess.domain.chessGame;
 
 import chess.domain.piece.Color;
 import chess.domain.piece.EmptyPiece;
