+++ src/main/java/chess/machine/Status.java
@@ -1,7 +1,7 @@
 package chess.machine;
 
-import chess.domain.chessBoard.BoardScore;
-import chess.domain.chessBoard.ChessGame;
+import chess.domain.chessGame.BoardScore;
+import chess.domain.chessGame.ChessGame;
 import chess.domain.piece.Color;
 import chess.view.OutputView;
